Move widget views to own files
